requirements.txt
PyYAML==5.1
requests==2.21.0
colour==0.1.5
